Full parser support for new state_<lang> snippet.
Full support for naming a piece of mutable state and giving it a
starting value for all AJVM languages. Compiler support and code
generation for Clojure only, Ruby and Scala next. Still need to add an
exception when the :state item of a :gen-class form is not Java-safe
(wish Clojure did this for us during compilation).